RAG Based Chatbot/code_Files/requirements.txt
groq
sentence-transformers
numpy
load_dotenv
pymongo
pdfplumber
docx2txt
